Updating in process

+++ molsysmt/tests/structure/get_distances/test_get_distances_from_molsysmt_MolSys.py
@@ -116,7 +116,7 @@ def test_get_distances_from_molsysmt_MolSys_10():
     n_structures = msm.get(molsys, n_structures=True)
     all_structure_indices = np.arange(n_structures)
     displacements = msm.structure.get_distances(molsys, selection="all",
-                             structure_indices=np.zeros(n_structures), structure_indices_2=all_structure_indices)
+                             structure_indices=np.zeros(n_structures, dtype=int), structure_indices_2=all_structure_indices)
     check_shape = ((5000,62,62)==displacements.shape)
     check_distance = np.isclose(puw.get_value(displacements[1000, 30, 30], to_unit='nm'), 0.4517681)
     assert check_shape and check_distance
